requirements.txt
streamlit==1.32.0
bcrypt==4.1.2
cryptography==42.0.5
PyYAML==6.0.1 

some install packages


Hussain 
12345678